refactor(recall): #250 の arm comparison を arms.rs に分割 (file-size gate 解消)
recall.rs が 473 行となり #255 の 400 行 file-size gate に抵触し CI test job が fail していた。#250 で追加した seed-stage Hit@k / query 分類 / arm comparison を src/recall/arms.rs に分離 (recall.rs 202 / arms.rs 287、いずれも gate 内)。pub use で recall:: 経由の参照を維持し count_u32 は pub(super) で共有。挙動・テスト不変 (recall:: 25 tests pass、clippy --all-features clean、fmt clean)。
